Web-Search-agent/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
lxml_html_clean
google-generativeai==0.3.2
python-dotenv==1.0.0
feedparser==6.0.10
lxml==4.9.3